Create welcomehomerobot
The raspberry pi will detect motion upon waking up in the morning.  It will log into an online weather service and tell you the current temperature, daily high and low, and the likelihood of rain.  Upon arriving home it will give you a welcome home message and then play a song.  
(Song files and voice files are not included here)